dash_test/requirements.txt
dash==2.17.1
pandas==2.2.2
plotly==5.22.0
python-dotenv==1.0.1
SQLAlchemy==2.0.31
